Refactor integration test suite & tools
