+++ crates/telos/node/tests/integration/mod.rs
@@ -1,0 +1,6 @@
+pub mod test_bad_memo;
+pub mod test_blocknum;
+pub mod test_nonce;
+pub mod test_resources_sandwich;
+pub mod test_revision;
+pub mod test_tx_types;
