+++ crates/telos/node/tests/main.rs
@@ -1,7 +1,4 @@
-// pub mod integration;
-// pub mod live_test_runner;
-// pub mod state_bypass;
-// pub mod storage_compare;
+pub mod integration;
 pub mod utils;
 
 const fn main() {}
